requirements.txt
flask
firebase-admin
numpy
pillow
opencv-python-headless
tensorflow
scikit-learn
gunicorn
